機能の細分化と細かい調整を行いました。
