+++ ServiceApplication1/Main.h
@@ -1,8 +1,13 @@
+//
+// Main.h
+//
+
 #pragma once
 
+#include "ServiceCommon.hpp"
 #include "ServiceCore.h"
+#include "ServiceCommand.h"
 #include "EventLogger.h"
-#include "ServiceMessage.h"
 
 VOID WINAPI SvcMain(DWORD dwArgc, LPTSTR* lptszArgv);
 VOID WINAPI CtrlHandler(DWORD dwControlCode);
