+++ ServiceApplication1/Service.cpp
@@ -12,7 +12,14 @@ Service::Service(EventLogger& logger)
 BOOL Service::Run()
 {
     _logger.TraceStart(CATEGORY_SERVICE, __FUNCTIONW__);
-    _logger.Log(EVENTLOG_SUCCESS, CATEGORY_SERVICE, SVC_SUCCESS_SYSTEM, 2, __FUNCTIONW__, _T("サービスを開始します。"));
+    
+    _logger.Log(
+        EVENTLOG_SUCCESS,
+        CATEGORY_SERVICE,
+        SVC_SUCCESS_SYSTEM,
+        2,
+        __FUNCTIONW__,
+        _T("サービスを開始します。"));
 
     while (TRUE)
     {
@@ -20,23 +27,48 @@ BOOL Service::Run()
         switch (dwRet)
         {
             case WAIT_TIMEOUT:
-                _logger.Log(EVENTLOG_INFORMATION_TYPE, CATEGORY_SERVICE, SVC_SUCCESS_API, 2, __FUNCTIONW__, _T("サービスを実行中です。"));
+                _logger.Log(
+                    EVENTLOG_INFORMATION_TYPE,
+                    CATEGORY_SERVICE,
+                    SVC_SUCCESS_API,
+                    2,
+                    __FUNCTIONW__,
+                    _T("サービスを実行中です。"));
                 break;
 
             case WAIT_OBJECT_0:
             case WAIT_ABANDONED_0:
-                _logger.Log(EVENTLOG_INFORMATION_TYPE, CATEGORY_SERVICE, SVC_SUCCESS_API, 2, __FUNCTIONW__, _T("サービス終了イベントを受信しました。"));
+                _logger.Log(
+                    EVENTLOG_INFORMATION_TYPE,
+                    CATEGORY_SERVICE,
+                    SVC_SUCCESS_API,
+                    2,
+                    __FUNCTIONW__,
+                    _T("サービス終了イベントを受信しました。"));
                 ::ExitThread(ERROR_SUCCESS);
                 break;
 
             default:
-                _logger.Log(EVENTLOG_ERROR_TYPE, CATEGORY_SERVICE, SVC_ERROR_API, 2, __FUNCTIONW__, _T("異常なサービス終了イベントを受信しました。"));
+                _logger.Log(
+                    EVENTLOG_ERROR_TYPE,
+                    CATEGORY_SERVICE,
+                    SVC_ERROR_API,
+                    2,
+                    __FUNCTIONW__,
+                    _T("異常なサービス終了イベントを受信しました。"));
                 ::ExitThread((DWORD)-1);
                 break;
         }
     }
 
-    _logger.Log(EVENTLOG_SUCCESS, CATEGORY_SERVICE, SVC_SUCCESS_SYSTEM, 2, __FUNCTIONW__, _T("サービスを終了します。"));
+    _logger.Log(
+        EVENTLOG_SUCCESS,
+        CATEGORY_SERVICE,
+        SVC_SUCCESS_SYSTEM,
+        2,
+        __FUNCTIONW__,
+        _T("サービスを終了します。"));
+
     _logger.TraceFinish(CATEGORY_SERVICE, __FUNCTIONW__);
     return TRUE;
 }
