+++ ServiceApplication1/ServiceApplication1.vcxproj
@@ -94,6 +94,7 @@
     </ClCompile>
     <ClCompile Include="Service.cpp" />
     <ClCompile Include="ServiceCommand.cpp" />
+    <ClCompile Include="ServiceCommon.cpp" />
     <ClCompile Include="ServiceCore.cpp" />
     <ClCompile Include="Thread.cpp" />
   </ItemGroup>
@@ -105,6 +106,7 @@
     <ClInclude Include="resource.h" />
     <ClInclude Include="Service.h" />
     <ClInclude Include="ServiceCommand.h" />
+    <ClInclude Include="ServiceCommon.hpp" />
     <ClInclude Include="ServiceCore.h" />
     <ClInclude Include="Thread.h" />
   </ItemGroup>
